phase3/object_detection/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
pandas>=1.3.0
numpy>=1.19.5
Pillow>=8.3.1
matplotlib>=3.4.3
tqdm>=4.62.2 